world: implement legacy chunk parsing

+++ lib/trailblazer/src/world/block.rs
@@ -1,4 +1,4 @@
 pub struct Block {
     id: u32,
-    shapes: Vec<[f32; 3]>
+    shapes: Vec<[f64; 3]>
 }
